CTW-1404 Refactoring data retrieval for encounters and documents (#1253)
Refactors data retrieval for encounters and documents so that we only
retrieve the last 20 when rendering the red dot. Also modifying the
encounters data retrieval to make it optional whether or not we also
retrieve all documents and use them to generate the clinical notes. That
said, still leaving much of that logic after the network requests
untouched.