Version Packages (#1257)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.68.5

### Patch Changes

- 70d85f63: Update care team component and only show completed
immunizations.
-   f6e870f2: Show clinical notes in document details drawer.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>