IFrame ZAP to pass same theme to CTWProvider as parent window. (#1276)
IFrame ZAP should have the same ThemeProvider context as the parent
window ThemeProvider. The `theme` passed into `CTWProvider` contains
both the tailwind theme and iframe theme. The
`ZusAggregatedProfileIFrame` component gets those two themes from the
`ThemeProvider` as separate objects. Originally we would send the iframe
theme as a separate configuration to the standalone ZAP, but then in the
standalone ZAP the `ThemeProvider` actually doesn't have the iframe
theme. The `Drawer` component now uses the `ThemeProvider` to access the
iframe theme as well so we need to make sure it's in the
`CTWProviderProps["theme"]` sent to the iframe.