Version Packages (#1308)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.69.8

### Patch Changes

- 5e386768: Temporarily remove hover actions in Hospitalizations table.
- c58d5e62: Fix hospitalization table's handling of renavigating to the
patient's page so it works with Remix.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>