Version Packages (#1334)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@2.0.0

### Major Changes

- 99226384: By default PatientConditionsAll will set
onlyAllowAddOutsideConditions to true. Previously this defaulted to
false.

### Patch Changes

- 99226384: Prevent PatientConditionsAll component from showing "add"
button in drawer when readOnly is true.
-   8f2b641f: Add functioning row actions to the ADT table.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>