CTW-1523 Remove practitioner query (#1335)
When a user clicks dismiss/read/etc, or when we generate a provenance
for data being written, we associate that data with the user via a
practitioner reference. If the user does not have a practitioner
reference then we just put a display reference with the user's email.

We have been including the practitioner's full name as a display
property on the reference which requires a call to the server to
retrieve the practitioner. This has been using a particularly slow ODS
query (up to 60s and a timeout in many cases).

This display property on an explicit practitioner reference is
unnecessary. It may make the data a bit easier to visually inspect
(without needing to look up the practitioner), but there's no real
use-case for us doing that right now and it seems like overkill to make
a request to the server to get this data right now.

At some point in the future we may need to look up the practitioner for
the current user (and when we do that we should use FQS), but for the
time being we should fix this performance issue by simply dropping the
display on the practitioner reference and stop calling the server for
this data.