Version Packages (#1340)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@2.0.3

### Patch Changes

-   c64f77e1: Remove row action functionality from ADT table due to bug.
-   824ff7cb: Remove zus-account header from package request

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>