Version Packages (#1351)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@2.1.1

### Patch Changes

-   12f4baea: Remove 3 month view from Hospitalizations
-   d3e7b9ae: Fix isLoading for patient history table

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>