CTW-1577 Fixing issue with showing add button when we shouldn't (#1362)
Jira Ticket Number: CTW-1577

## What does this PR accomplish?

We have been showing the "Add" button in the medication drawer even when
`readOnly` is set to true. This PR fixes that issue.

## How did you test it?

Tested locally with `readOnly` set both ways.

## How will you know that this is working after deployment?

I will verify in Elation sandbox that it is set up as expected.

## Any outstanding items or follow ups with Jira tickets?

Need to do CTW version bump.