Ctw 1356/request history last refresh (#1365)
Jira Ticket Number: https://zeushealth.atlassian.net/browse/CTW-1356

## What does this PR accomplish?

Fixes a bug showed no requested history even though a patient had data.
Essentially when patients had multiple patient records we at random
chose one FHIR id and use that as the one we used to hit the patient
history API. This behavior is incorrect as we may have pulled the data
under one of the other patient id's. We now look at at all the fhir id's
under a patient. This is accomplished by iterating over all the patient
id's for a given patient.

## How did you test it?

I tested this by navigating over to the patient in question mentioned in
the bug report and verifying that we indeed now show a requested date.