Favor showing RxNorm display values for medication (#1367)
Improve how we find the appropriate "label" for a codeable concept. We
can now optionally specify a system to favor, in which case we'll try
and find the `display` value associated with that system coding
(skipping those with extensions so that we skip enrichments).
Additionally, instead of falling back to the first coding when there is
no text, we now instead fall back to the first coding with a display,
and then first coding with a code.