Version Packages (#1381)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@2.3.0

### Minor Changes

- f9b0ee64: Upgrade typescript, i18next, react-i18next and lodash types.

### Patch Changes

- f9aea808: Add patch to remove a polyfill that may be causing auth
issues for CTW stand alone.
- aa860c65: Provides a method of detecting how many ZAP tabs have
notifications.
-   cd17dce2: Minor CSS fix for add filters in CTW

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>