[Bug] Fix Patient Search when patients have the same name (#1396)
Jira Ticket Number:
[CTW-1620](https://zeushealth.atlassian.net/browse/CTW-1620)

This PR fixes a bug in patient search when there are multiple patients
with the same name. Previously the `ComboBoxOption`'s value was the
label, which comprised the patient's full name. When an option was
selected, the `ComboBoxField` would use the label to look up the first
occurrence in the results, so if there are 2 or more patients with the
same full name, it would always return the first patient.

Now the `ComboBoxOption`'s value is the entire option so we can just
return the option's value to the `onCustomSelectChange` callback.

## How did you test it?

Manually tested in the demo app.
Manually tested with relative dependancies in standalone.

## How will you know that this is working after deployment?

Manually test in all environments


[CTW-1620]:
https://zeushealth.atlassian.net/browse/CTW-1620?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ