Version Packages (#1397)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@2.3.5

### Patch Changes

- 72ef7320: Fix ComboBoxField to select the correct option when multiple
options have identical labels

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>